requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
mysqlclient>=2.2